requirements.txt
elevenlabs
python-dotenv==0.21.0
